Tweak HSync behavior
Amstrad used 14us, but the PAL and NTSC standards are
closer to 12us, with 52us of active line.

Apparently the Gate Array runs the actual sync signal
for 6us anyway, so this is closer to a blank signal
than a sync signal.

Move sync timing accordingly: since the IRQ comes from
the end of the CRTC's HSYNC, shortening HSYNC by 2us
changes the IRQ and HALT timing by as much